+++ src/corpus/config/CorpusConfigCustom.vue
@@ -9,6 +9,7 @@ import LayoutBox from "@/components/LayoutBox.vue";
 import type { MinkResponse } from "@/api/api.types";
 import useMessenger from "@/message/messenger.composable";
 import SyntaxHighlight from "@/components/SyntaxHighlight.vue";
+import PendingContent from "@/spin/PendingContent.vue";
 
 const corpusId = useCorpusIdParam();
 const { config, uploadConfigRaw } = useConfig(corpusId);
@@ -46,12 +47,16 @@ async function upload(files: File[]) {
 
   <div class="flex flex-wrap gap-4">
     <LayoutBox class="w-96 grow" :title="$t('show')">
-      <SyntaxHighlight v-if="config" language="yaml" :code="config" />
+      <PendingContent :on="`corpus/${corpusId}/config`">
+        <SyntaxHighlight v-if="config" language="yaml" :code="config" />
+      </PendingContent>
     </LayoutBox>
 
     <LayoutBox class="w-96 grow" :title="$t('upload')">
       <HelpBox>{{ $t("config.custom.upload.help") }}</HelpBox>
-      <FileUpload :file-handler="upload" accept=".yaml,.yml" />
+      <PendingContent :on="`corpus/${corpusId}/config`" blocking>
+        <FileUpload :file-handler="upload" accept=".yaml,.yml" />
+      </PendingContent>
     </LayoutBox>
   </div>
 </template>