+++ src/main.ts
@@ -18,7 +18,8 @@ const app = createApp(App) //
 // Use the Matomo plugin only if configured in env.
 if (import.meta.env.VITE_MATOMO_URL && import.meta.env.VITE_MATOMO_ID) {
   app.use(matomo, {
-    host: import.meta.env.VITE_MATOMO_URL,
+    // Url expected without trailing slash
+    host: import.meta.env.VITE_MATOMO_URL.replace(/\/$/, ""),
     siteId: import.meta.env.VITE_MATOMO_ID,
     router,
   });