+++ CHANGELOG.md
@@ -13,6 +13,10 @@ As this project is a user-facing application, the places in the semantic version
 
 ## [Unreleased]
 
+### Fixed
+
+- Show loading spinners on the custom config page
+
 ## [1.6.1] (2024-08-13)
 
 ### Added