+++ CHANGELOG.md
@@ -17,6 +17,7 @@ As this project is a user-facing application, the places in the semantic version
 
 - Show loading spinners on the custom config page
 - Strip slash from Matomo url
+- Added `.editorconfig`
 
 ## [1.6.1] (2024-08-13)