+++ CHANGELOG.md
@@ -17,6 +17,10 @@ As this project is a user-facing application, the places in the semantic version
 
 - Track some events to Matomo [#166](https://github.com/spraakbanken/mink-frontend/issues/166)
 
+### Changed
+
+- Improved repo documentation
+
 ### Fixed
 
 - Show loading spinners on the custom config page